Re-organized exceptions module
