+++ pyutils/core/exceptions.py
